backend/requirements.txt
lmdb
paho-mqtt
zabbix-api # https://pypi.org/project/zabbix-api/
